+++ test/config-webpack-serve.spec.js
@@ -10,22 +10,22 @@ describe('config webpack serve', () => {
 
   const skyuxConfig = {
     runtime: runtimeUtils.getDefaultRuntime(),
-    skyux: {
-      mode: '',
-      host: {
-        url: 'https://my-host-server.url'
-      },
+    skyux: runtimeUtils.getDefaultSkyux({
       app: {
         externals: {
           test: true
         }
+      },
+      host: {
+        url: 'https://my-host-server.url'
       }
-    }
+    })
   };
 
   let lib;
   let called;
-  let openCalledWith;
+  let openParamUrl;
+  let openParamBrowser;
   let config;
   let argv = {};
 
@@ -42,13 +42,16 @@ describe('config webpack serve', () => {
 
   beforeEach(() => {
     called = false;
-    openCalledWith = '';
+    openParamUrl = '';
+    openParamBrowser = undefined;
+
     argv = {};
 
     spyOn(logger, 'info');
-    mock('open', (url) => {
+    mock('open', (url, browser) => {
       called = true;
-      openCalledWith = url;
+      openParamUrl = url;
+      openParamBrowser = browser;
     });
 
     lib = require('../config/webpack/serve.webpack.config');
@@ -129,10 +132,9 @@ describe('config webpack serve', () => {
       }
     });
 
+    const url = 'https://my-host-server.url/@blackbaud/skyux-builder/?local=true&_cfg=';
     expect(logger.info).toHaveBeenCalledTimes(2);
-    expect(openCalledWith).toContain(
-      'https://my-host-server.url/@blackbaud/skyux-builder/?local=true&_cfg='
-    );
+    expect(openParamUrl.indexOf(url)).toBe(0);
   });
 
   it('should log the host url and launch it when --launch host', () => {
@@ -154,8 +156,9 @@ describe('config webpack serve', () => {
       }
     });
 
+    const url = 'https://my-host-server.url';
     expect(logger.info).toHaveBeenCalledTimes(2);
-    expect(openCalledWith).toContain('https://my-host-server.url');
+    expect(openParamUrl.indexOf(url)).toBe(0);
   });
 
   it('should log the local url and launch it when --launch local', () => {
@@ -177,8 +180,9 @@ describe('config webpack serve', () => {
       }
     });
 
+    const url = 'https://localhost:1234';
     expect(logger.info).toHaveBeenCalledTimes(2);
-    expect(openCalledWith).toContain('https://localhost:1234');
+    expect(openParamUrl.indexOf(url)).toBe(0);
   });
 
   it('should log a done message and not launch it when --launch none', () => {
@@ -227,14 +231,12 @@ describe('config webpack serve', () => {
     expect(called).toEqual(false);
   });
 
+
+
   it('host querystring should not contain externals if they do not exist', () => {
     const localConfig = lib.getWebpackConfig(argv, {
       runtime: runtimeUtils.getDefaultRuntime(),
-      skyux: {
-        host: {
-          url: ''
-        }
-      }
+      skyux: runtimeUtils.getDefaultSkyux()
     });
 
     localConfig.plugins.forEach(plugin => {
@@ -255,7 +257,7 @@ describe('config webpack serve', () => {
                   chunks: []
                 })
               });
-              const urlParsed = urlLibrary.parse(openCalledWith, true);
+              const urlParsed = urlLibrary.parse(openParamUrl, true);
               const configString = new Buffer.from(urlParsed.query._cfg, 'base64').toString();
               const configObject = JSON.parse(configString);
 
@@ -291,13 +293,14 @@ describe('config webpack serve', () => {
                   ]
                 })
               });
-              const urlParsed = urlLibrary.parse(openCalledWith, true);
+              const urlParsed = urlLibrary.parse(openParamUrl, true);
               const configString = new Buffer.from(urlParsed.query._cfg, 'base64').toString();
               const configObject = JSON.parse(configString);
+              const url = 'https://localhost:1234';
 
               expect(urlParsed.query._cfg).toBeDefined();
               expect(configObject.externals).toEqual(skyuxConfig.skyux.app.externals);
-              expect(configObject.localUrl).toContain('https://localhost:1234');
+              expect(configObject.localUrl.indexOf(url)).toBe(0);
               expect(configObject.scripts).toEqual([
                 { name: 'a.js' },
                 { name: 'b.js' }
@@ -313,14 +316,14 @@ describe('config webpack serve', () => {
     argv.envid = 'asdf';
 
     bindToDone();
-    expect(openCalledWith).toContain(`?envid=asdf`);
+    expect(openParamUrl).toContain(`?envid=asdf`);
   });
 
   it('should pass through svcid from the command line', () => {
     argv.svcid = 'asdf';
 
     bindToDone();
-    expect(openCalledWith).toContain(`?svcid=asdf`);
+    expect(openParamUrl).toContain(`?svcid=asdf`);
   });
 
   it('should run envid and svcid through encodeURIComponent', () => {
@@ -328,7 +331,7 @@ describe('config webpack serve', () => {
     argv.svcid = '^%';
 
     bindToDone();
-    expect(openCalledWith).toContain(
+    expect(openParamUrl).toContain(
       `?envid=${encodeURIComponent(argv.envid)}&svcid=${encodeURIComponent(argv.svcid)}`
     );
   });
@@ -339,8 +342,21 @@ describe('config webpack serve', () => {
     argv.myid = 'asdf3';
 
     bindToDone();
-    expect(openCalledWith).toContain(`?envid=asdf1&svcid=asdf2`);
-    expect(openCalledWith).not.toContain(`myid=asdf3`);
+    expect(openParamUrl).toContain(`?envid=asdf1&svcid=asdf2`);
+    expect(openParamUrl).not.toContain(`myid=asdf3`);
+  });
+
+  it('should pass --browser flag to open', () => {
+    argv.browser = 'custom-browser';
+    bindToDone();
+    expect(openParamBrowser).toEqual(argv.browser);
+  });
+
+  it('should handle --browser edge different syntax', () => {
+    argv.browser = 'edge';
+    bindToDone();
+    expect(openParamBrowser).not.toBeDefined();
+    expect(openParamUrl.indexOf('microsoft-edge')).toBe(0);
   });
 
 });